sort product endpoint and new items
