+++ client/src/components/cards/index.js
@@ -1,0 +1,3 @@
+export { default as AdminProductCard } from './AdminProductCard'
+export { default as Jumbotron } from './Jumbotron'
+export {default as ProductCard} from './ProductCard'
